Added new filter based on LND dataset
Words that apperead less than 3 times in LND dataset were removed. The same process will be done with LAS (first we need to add the headlines).  On the cluster, headlines are being added to LAS, then we will get the final lexicon (words not appearing in LAS will be removed, it has do be done again because for some reason the results are different from last time). From there, rows have to be removed from LAS according to LND_lexicon. Meanwhile I will compute LND embeddings, when LAS is over we will have to clean the rows using LLMs and finally compute the embeddings for LAS too.